requirements.txt
flask
flask-sijax
black # 代码格式化
load_dotenv # 加载环境变量
requests-toolbelt # 可以处理上传文件
apscheduler # 定时任务管理的库